second commit

+++ src/main/java/tn/spring/springboot/Repository/EtudiantRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import tn.spring.springboot.Entities.Etudiant;
 
 //repository that extends CrudRepository
-public interface EtudiantRepository extends CrudRepository<Etudiant, Integer>
+public interface EtudiantRepository extends CrudRepository<Etudiant,Long>
 {
 }
